requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
python-dateutil==2.8.1
six==1.15.0
soupsieve==2.2 